requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-chroma
langgraph

google-genai
tavily-python

chromadb
sentence-transformers

gradio
requests
beautifulsoup4
python-dotenv

openai
ollama
rich
crewai